[brain tests] Burst attr from the main file
